requirements.txt
streamlit
pypdf
sentence-transformers
numpy
openpyxl
groq
python-dotenv